requirements.txt
cachetools==4.2.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
et-xmlfile==1.1.0
Flask==1.1.2
future==0.18.2
google-api-core==1.34.0
google-auth==1.35.0
google-cloud-core==1.7.3
google-cloud-firestore==1.6.0
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.48.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.2.2
openpyxl==3.1.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
python-dotenv==0.14.0
pytz==2022.7.1
requests==2.24.0
requests-oauthlib==1.0.0
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
urllib3==1.25.10
webexteamssdk==1.6
Werkzeug==1.0.1
zipfile38==0.0.3